+++ src/components/ES5/simple-3D-mouse-interaction/js/3D-mouse-interaction.js
@@ -157,7 +157,7 @@ APP = ( function ( APP, $, window, document ) {
 
 			}
 
-			
+			//update camera and controls
 			controls.update();
 			
 			renderer.render( scene, camera );